requirements.txt
Flask==2.1.3
mysql-connector-python==8.0.30